requirements.txt
streamlit>=1.32.0
plotly>=5.18.0
pandas>=2.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
numpy>=1.26.0
scipy>=1.12.0
